adds so the ean13 and upc accepts integer value numbers
